+++ tests/Support/AllControllers/MoreController/Nested/MyController.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Yii\View\Renderer\Tests\Support\AllControllers\MoreController\Nested;
+
+final class MyController
+{
+}